social_network_backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-cors-headers==4.3.1
django-log-viewer==1.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==20.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
